% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brgeojson-data.R
\docType{data}
\name{brgeojson}
\alias{brgeojson}
\title{Geojson data for plotly}
\format{
A json file
}
\usage{
data(brgeojson)
}
\description{
This is a json file with geoposition information for integrative maps
}
\examples{
data(brgeojson)
}
\keyword{geojson}
