% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr-data.R
\name{cbr_data}
\alias{cbr_data}
\title{Read and bind multiple files of COVID-19 Portal}
\usage{
cbr_data(dir = ".", filepath = ".", save2disk = FALSE)
}
\arguments{
\item{dir}{a \code{character} string, corresponding the directory containing the
files downloaded from the COVID-19 Portal.}

\item{filepath}{a \code{character} string, corresponding the folder where files will
be save. By default, the current directory is used
(see \code{\link[=getwd]{getwd()}}).}

\item{save2disk}{\code{logical}. If \link{TRUE}, the database is saved in a file in
csv format.}
}
\value{
a \code{data.table} object with 17 columns, corresponding the database.
}
\description{
This function read and bind all files available in a directory
and returns the database in data table format.
}
\examples{
\dontrun{
covid_br_db <- cbr_data("./data-raw")
}
}
