% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr-download.R
\name{cbr_download}
\alias{cbr_download}
\title{Download a file from an existing download URL}
\usage{
cbr_download(
  url_file = NULL,
  dir = getwd(),
  filename = "covidBR",
  keep_filename = FALSE,
  overwrite = TRUE,
  show_progress = getOption("covidBR.verbose", default = interactive())
)
}
\arguments{
\item{url_file}{a \code{character} string, corresponding the URL to download the
public file.}

\item{dir}{a \code{character} string, corresponding the directory where file
should be downloaded. By default, the current directory is used
(see \code{\link[=getwd]{getwd()}}).}

\item{filename}{a \code{character} string, corresponding the filename (without the
extension) saved
on the local disc. By default, \strong{covidBR} was used as filename.}

\item{keep_filename}{logical. If \link{TRUE}, the name of public
file will be used and parameter \code{filename} is ignored (\code{default = FALSE}).}

\item{overwrite}{\code{logical}. If \link{TRUE}, the default, the local file with the
#' same name will be overwritten.}

\item{show_progress}{\code{logical}. If \link{TRUE}, the default, shows a progress bar
for uploading.}
}
\description{
Download a public file from an existing download URL.
}
\examples{
\dontrun{
# Get dowload url (COVID-19 Portal)
dowload_url <- cbr_dowload_url()

# Save file with default name ("data")
cbr_download(url_file = dowload_url)

# Save file with name "my_data"
cbr_download(url_file = dowload_url, filename = "mydata")

# Save file with default name of COVID-19 Portal
cbr_download(url_file = dowload_url, keep_filename = TRUE)

# Save file in 'myfolder' folder
cbr_download(url_file = dowload_url, dir = "./myfolder")

}
}
