% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr-download-url.R
\name{cbr_download_url}
\alias{cbr_download_url}
\title{Get the download url of a public file}
\usage{
cbr_download_url()
}
\value{
the URL to download the public file.
}
\description{
The original data comes from the Brazilian
COVID-19 Portal (Portal do COVID-19), and can be found at
\url{https://covid.saude.gov.br}. Annually, the Ministry of Health
(Ministerio da Saude) updates and publicly disseminates the database, and,
consequently, the download url is updated. This function scrape the link
address of the web page and returns the URL download for the public file.
}
\examples{
 \dontrun{cbr_dowload_url()}

}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:content]{httr::content()}}
}
