% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr-unzip.R
\name{cbr_unzip}
\alias{cbr_unzip}
\title{Extract contents of an archive}
\usage{
cbr_unzip(zip_file, dir = ".", keep_zip_file = TRUE)
}
\arguments{
\item{zip_file}{a \code{character} string, corresponding the archive filename, or
an \link{archive} object.}

\item{dir}{a \code{character} string, corresponding the directory to extract
archive contents. By default, the current directory is used.
(see \code{\link[=getwd]{getwd()}}).}

\item{keep_zip_file}{logical. If \link{TRUE}, deletes the zip archive.}
}
\description{
Extract contents of an archive to a directory.
}
\examples{
\dontrun{
cbr_unzip(file.choose())
}
}
\seealso{
\code{\link[archive:archive_extract]{archive::archive_extract()}},\code{\link[base:unlink]{base::unlink()}}
}
