% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr-update-date.R
\name{cbr_update_date}
\alias{cbr_update_date}
\title{Get the update date}
\usage{
cbr_update_date()
}
\value{
a \code{character} string, corresponding to the date when the public
files were last updated.
}
\description{
This function scrape the link
address of the web page and returns the update date of the public files.
}
\examples{
\dontrun{cbr_update_date()}
}
