% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{createMaps}
\alias{createMaps}
\title{Create maps with database of covidBR}
\usage{
createMaps(dir = "./", date = NULL)
}
\arguments{
\item{dir}{of your database generated with downloadCovidBR function \code{directory}}

\item{date}{for maps, if NULL, is the last \code{date()}}
}
\value{
\code{message()}
}
\description{
Create maps with database of covidBR
}
\author{
Guilherme Fernandes Castro de Oliveira
}
