% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadCovidBR}
\alias{downloadCovidBR}
\title{Download data of covid in brasil}
\usage{
downloadCovidBR(dir = "./", maps = FALSE, date = NULL, withCorr = FALSE)
}
\arguments{
\item{dir}{to save data \code{directory}}

\item{maps}{with maps or no \code{boolean()}}

\item{date}{for maps, if NULL, is the last \code{date()}}

\item{withCorr}{for maps, with correlation of data or no \code{boolean()}}
}
\value{
\code{message()}
}
\description{
Download data of covid in brasil
}
\author{
Guilherme Fernandes Castro de Oliveira
}
