% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibgeids.R
\name{ibge.id}
\alias{ibge.id}
\title{Find IBGE code}
\usage{
ibge.id(location)
}
\arguments{
\item{location}{string (state or city/county name)}
}
\value{
A single line data frame with IBGE code, location name and the estimated population
}
\description{
The function returns the IBGE code based on state (ex: 'SP', 'MG' ...) or the city name (ex: 'Rio de Janeiro', São Paulo'...)
}
\examples{
myId <- ibge.id('SP')

}
