% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge-data.R
\docType{data}
\name{ibge}
\alias{ibge}
\title{IBGE data}
\format{
data frame
}
\usage{
data(ibge)
}
\description{
This is data frame with IBGE information for covidBR use
}
\examples{
data(ibge)
}
\keyword{ibge}
