% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadata.R
\name{load.covidBr}
\alias{load.covidBr}
\title{Load COVID19 data}
\usage{
load.covidBr()
}
\value{
A data frame with covid19 information
}
\description{
The function downloads the updated COVID19 information from Brazil
}
\examples{
covid <- load.covid()
head(covid)
}
