% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadibge.R
\name{load.ibgeinfo}
\alias{load.ibgeinfo}
\title{Load IBGE data}
\usage{
load.ibgeinfo()
}
\value{
A sorted data frame with IBGE codes, location names and and estimated population
}
\description{
The function downloads IBGE information from Brazil
}
\examples{
ibge <- load.ibgeinfo()
}
