% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casesmap.R
\name{view.cases.map}
\alias{view.cases.map}
\title{Brazil COVID19  map - new cases}
\usage{
view.cases.map(x, perhab = FALSE)
}
\arguments{
\item{x}{a data frame with COVID19 information from load.covid()}

\item{perhab}{boolean (default = FALSE). True displays the incidence per 1 million people.}
}
\value{
A plotly map from Brazil with covid19 incidence per state
}
\description{
The function creates line chart displaying the number of new COVID19 cases (7-days mean incidence) in Brazil
}
\examples{
covid <- load.covidBr()
mymap <- view.cases.map(covid)

}
