% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdeath.R
\name{view.covid.deaths}
\alias{view.covid.deaths}
\title{Plot COVID19 deaths}
\usage{
view.covid.deaths(x, ibge = NA, top5 = FALSE)
}
\arguments{
\item{x}{a data frame with COVID19 information from load.covid()}

\item{ibge}{a data frame (default = NA, all Brazil data) from ibge.id()}

\item{top5}{boolean (default = FALSE). True if id == NA. If True, displays the top5 states in total number of cases}
}
\value{
A line chart with covid19 deaths in function of time
}
\description{
The function creates line chart displaying the number of COVID19 deaths (7-days mean) in Brazil
}
\examples{
covid <- load.covidBr()
ibge <- load.ibgeinfo()
myid <- ibge.id(ibge, 'SP')
mychart <- view.covid.deaths(covid, myid)

}
