% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewnew.R
\name{view.covid.newcases}
\alias{view.covid.newcases}
\title{Plot COVID19 new cases}
\usage{
view.covid.newcases(x, ibgeinfo = NA, top5 = FALSE)
}
\arguments{
\item{x}{a data frame with COVID19 information from load.covid()}

\item{ibgeinfo}{a data frame (default = NA, all Brazil data) from ibge.id()}

\item{top5}{boolean (default = FALSE). True if id == NA. If True, displays the top5 states in total number of cases}
}
\value{
A line chart with covid19 incidence in function of time
}
\description{
The function creates line chart displaying the number of new COVID19 cases (7-days mean incidence) in Brazil
}
\examples{
covid <- load.covidBr()
myid <- ibge.id('SP')
mychart <- view.covid.newcases(covid, myid)

}
