% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathsmap.R
\name{view.deaths.map}
\alias{view.deaths.map}
\title{Brazil COVID19  map - deaths}
\usage{
view.deaths.map(x, perhab = FALSE)
}
\arguments{
\item{x}{a data frame with COVID19 information from load.covid()}

\item{perhab}{boolean (default = FALSE). True displays the number of deaths per million people.}
}
\value{
A plotly map from Brazil with covid19 deaths per state
}
\description{
The function creates a interactive map displaying the total number number of COVID19 deaths in Brazil
}
\examples{
covid <- load.covidBr()
mymap <- view.deaths.map(covid)

}
