% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-maps.R
\name{createMaps}
\alias{createMaps}
\title{Create maps with database of covidBR}
\usage{
createMaps(dir = "./", date = NULL, withCorr = FALSE)
}
\arguments{
\item{dir}{of your database generated with downloadCovidBR function \code{directory}}

\item{date}{for maps, if NULL, is the last \code{date()}}

\item{withCorr}{make a correlation with data or no \code{boolean()}}
}
\value{
\code{message()}
}
\description{
Create maps with database of covidBR
}
\author{
Guilherme Fernandes Castro de Oliveira
}
