% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadCovidBR}
\alias{downloadCovidBR}
\title{Download data of covid in brasil}
\usage{
downloadCovidBR(dir = "./")
}
\arguments{
\item{dir}{to save data}
}
\description{
Download data of covid in brasil
}
\author{
Guilherme Fernandes Castro de Oliveira
}
